refactor: cn 함수 utils -> lib 폴더 이전
