+++ web/src/components/analyze/FirstResultCard.tsx
@@ -1,8 +1,8 @@
+import { cn } from "@/lib/cn";
 import type {
   AnalysisFirstCardData,
   ResultCardStyle,
 } from "@/types/analyzeResult/analyzeResult.types";
-import { cn } from "@/utils/cn";
 
 import { ANALYSIS_STATUS, STATUS_TEXT_MAP } from "@/constants/analyze/result";
