+++ web/src/components/analyze/TabSwitcher.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from "react";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/cn";
 
 export type TabCategory = "url" | "case";
