+++ web/src/components/common/BottomFullButton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/cn";
 
 export interface BottomFullButtonProps {
   state: boolean;
