+++ web/src/components/common/BottomSignupButton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/cn";
 
 export interface BottomSignupButtonProps {
   state: boolean;
