+++ web/src/components/common/NameTag.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/cn";
 
 interface NameTagProps {
   name: string;
