+++ web/src/components/communityFeed/CommunityFeedTab.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/cn";
 
 import { COMMUNITY_FEED_TABS } from "@/constants/commnityFeedTabs";
