+++ web/src/components/communityFeed/CommunityPostPreview.tsx
@@ -2,8 +2,8 @@ import { useNavigate } from "react-router-dom";
 
 import { path } from "@/routes/path";
 
+import { cn } from "@/lib/cn";
 import type { CommunityPost } from "@/types/community/community.types";
-import { cn } from "@/utils/cn";
 import { formatUTCtoKR } from "@/utils/formatUTCtoKR";
 
 import { ImageWithLoader } from "@/components/common/ImageWithLoader";
