+++ web/src/components/my/AnalysisLevelBadge.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/cn";
 
 type AnalysisStatus = "SAFE" | "WARNING" | "DANGER";
