+++ web/src/components/my/MyMenuItem.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode } from "react";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/cn";
 
 import { Toggle } from "@/components/common/Toggle";
